refactor(act): move iterate into internal/event-sourcing, split csv into its own module
\`iterate\` was sitting in the public \`transfer.ts\` module alongside
\`CsvFile\`, but it has a single caller — \`scan\` — and nothing
external uses it directly. It's an implementation detail of the
scan loop, not a public surface. Moves to
\`libs/act/src/internal/event-sourcing.ts\` next to its consumer and
drops out of the public barrel.

The CSV file adapter stays public (the inspector and any operator
tooling need to instantiate it), but lands in its own \`csv.ts\` —
\`transfer.ts\` is gone. The public surface for ACT-1128 is now just
\`CsvFile\` + \`EventSource\` + \`EventSink\`.

The csv spec file (\`libs/act/test/csv.spec.ts\`) imports \`iterate\`
via \`../src/internal/event-sourcing.js\` for the backpressure +
mailbox tests, matching the deep-import pattern \`restore.spec.ts\`
uses for \`scan\`. Public CsvFile tests use the regular import.

Coverage: 100% statements / 100% branches / 100% functions / 100% lines.
