+++ libs/act/src/index.ts
@@ -9,7 +9,7 @@ export * from "./act.js";
 export * from "./adapters/index.js";
 export * from "./builders/index.js";
 export * from "./config.js";
+export * from "./csv.js";
 export * from "./ports.js";
-export * from "./transfer.js";
 export * from "./types/index.js";
 export * from "./utils.js";
